+++ src/inmanta/ast/constraint/expression.py
@@ -18,7 +18,7 @@
 
 import re
 from abc import ABCMeta, abstractmethod
-from typing import Dict
+from typing import Dict, Optional
 
 from inmanta.ast import LocatableString, TypingException
 from inmanta.ast.statements import Literal, ReferenceStatement
@@ -62,9 +62,13 @@ class OpMetaClass(ABCMeta):
 
 
 class IsDefined(ReferenceStatement):
-    def __init__(self, attr: Reference, name: LocatableString) -> None:
-        super(IsDefined, self).__init__([attr])
-        self.attr = attr.root_in_self()
+    def __init__(self, attr: Optional[Reference], name: LocatableString) -> None:
+        if attr:
+            children = [attr]
+        else:
+            children = []
+        super(IsDefined, self).__init__(children)
+        self.attr = attr
         self.name = str(name)
 
     def requires_emit(self, resolver: Resolver, queue: QueueScheduler) -> Dict[object, ResultVariable]:
@@ -76,7 +80,10 @@ class IsDefined(ReferenceStatement):
         self.copy_location(resumer)
 
         # wait for the instance
-        RawUnit(queue, resolver, self.attr.requires_emit(resolver, queue), resumer)
+        if self.requires():
+            RawUnit(queue, resolver, super().requires_emit(resolver, queue), resumer)
+        else:
+            resumer.resume({}, resolver, queue)
         return {self: temp}
 
     def execute(self, requires: Dict[object, object], resolver: Resolver, queue: QueueScheduler) -> object:
@@ -84,10 +91,10 @@ class IsDefined(ReferenceStatement):
         return requires[self]
 
     def pretty_print(self) -> str:
-        name = "%s.%s is defined" % (self.attr, self.name)
-        if name[: len("self.")] == "self.":
-            name = name[len("self.") :]
-
+        if self.attr is not None:
+            name = "%s.%s" % (self.attr.pretty_print(), self.name)
+        else:
+            name = self.name
         return "%s is defined" % name